Поправил ошибки

+++ app/Http/Controllers/Admin/TaskRequestController.php
@@ -56,7 +56,7 @@ class TaskRequestController extends Controller
             'taskRequests' => $taskRequests,
             'filters' => $filters,
             'availableFilters' => [
-                'customers' => $this->taskService->getAvailableFilters(['taskRequestCustomers'], (int)Auth::id()),
+                'customers' => $this->taskService->getAvailableFilters(['taskRequestCustomers'], Auth::id()),
             ],
         ]);
     }
